Fix spurious warning about Fortran ordering
This warning is trying to ensure C-contiguity of the input.
However, F-contiguity and C-contiguity are neither exclusive nor comprehensive.
It's better to simply check directly if the array is C-contiguous.

Example when a dimension has length 1 so that an array is simultaneously F- and C-contiguous, and would lead to a spurious warning:

```python
m = np.array([[1, 2]])
assert m.flags.c_contiguous and m.flags.f_contiguous
```